move playlist logic to the extension
